+++ taskgraph/Task.py
@@ -23,7 +23,7 @@ import retrying
 __version__ = get_distribution('taskgraph').version
 
 
-_VALID_PATH_TYPES = (str, pathlib.Path)
+_VALID_PATH_TYPES = (str, pathlib.PurePath)
 _TASKGRAPH_DATABASE_FILENAME = 'taskgraph_data.db'
 
 try:
@@ -715,6 +715,7 @@ class TaskGraph(object):
                     LOGGER.warning(
                         "A duplicate task was submitted: %s original: %s",
                         new_task, self._task_hash_map[new_task])
+                    self._added_task_count -= 1
                     return duplicate_task
                 disjoint_target_set = (
                     new_task_target_set.symmetric_difference(